% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mosaicModule.R
\name{mod_mosaicModule_server}
\alias{mod_mosaicModule_server}
\title{Shiny Module for mosaicplot}
\usage{
mod_mosaicModule_server(id, inputData)
}
\arguments{
\item{id}{if of module}

\item{inputData}{"reactive" data}
}
\description{
mosaicModule Server Functions
}
