% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mosaicModule.R
\name{mod_mosaicModule_ui}
\alias{mod_mosaicModule_ui}
\title{Shiny Module for mosaicplot}
\usage{
mod_mosaicModule_ui(id)
}
\arguments{
\item{id}{id of module}
}
\description{
mosaicModule UI Function
}
