% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pairModule.R
\name{mod_pairModule_server}
\alias{mod_pairModule_server}
\title{Shiny Module for pair plot}
\usage{
mod_pairModule_server(id, inputData)
}
\arguments{
\item{id}{if of module}

\item{inputData}{"reactive" data}
}
\description{
PairModule Server Functions
}
