% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGen.R
\name{plotGen}
\alias{plotGen}
\title{plotGen}
\usage{
plotGen(
  data,
  type,
  criteria,
  describe = NULL,
  options = NULL,
  wraptype = NULL,
  wrapcols = NULL,
  wraprows = NULL
)
}
\arguments{
\item{data}{Data to describe in graph}

\item{type}{Type of graph}

\item{criteria}{X-axis variable in data}

\item{describe}{Y-axis variable in data}

\item{options}{Additional Options}
}
\value{
A \code{plotly} graph.
}
\description{
plotGen
}
