% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapVisModule.R
\name{mod_mapVisModule_ui}
\alias{mod_mapVisModule_ui}
\title{SHiny Module for map visualization.}
\usage{
mod_mapVisModule_ui(id, i18n = NULL)
}
\arguments{
\item{id}{id of module}

\item{i18n}{shiny.i18n object, default value is NULL}
}
\description{
mapVisModule UI Function
}
